+++ examples/Annotations/AnnotatedBench.php
@@ -54,7 +54,7 @@ class AnnotatedBench
      * @Assert("mode(variant.time.avg) < 200 ms")
 // endsection: assert
 // section: format
-     * @Format("mode(variant.time.avg) as ms ~ \" Hello World\"")
+     * @Format("mode(variant.time.avg) as ms ~ 'Hello World'")
 // endsection: format
 // section: executor
      * @Executor("local")