+++ lib/Benchmark/Runner.php
@@ -254,6 +254,10 @@ final class Runner
 
         // create the variants.
         foreach ($paramsIterator as $parameterSet) {
+            if (false === $parameterSet->nameMatches($config->getVariantFilters())) {
+                continue;
+            }
+
             foreach ($subjectMetadata->getIterations() as $nbIterations) {
                 foreach ($subjectMetadata->getRevs() as $revolutions) {
                     foreach ($subjectMetadata->getWarmup() as $warmup) {